Refactor registration process by moving signup logic to a separate module and simplifying form handling
